Former-commit-id: 85cb06da25cd4f1ba9abe726ace8d2e454ee5a88

+++ Sources/Internal/Base/FixedSizePoolAllocator.cpp
@@ -159,7 +159,7 @@ bool FixedSizePoolAllocator::CheckIsPointerValid(void * blockvoid)
         if ((block >= (uint8*)currentAllocatedBlockArray + sizeof(uint8*)) && (block < (uint8*)currentAllocatedBlockArray + sizeof(uint8*) + blockSize * blockArraySize))
         {
             // we are inside check is block correct.
-            uint32 shift = block - ((uint8*)currentAllocatedBlockArray + sizeof(uint8*));
+            uint32 shift = static_cast<uint32>(block - ((uint8*)currentAllocatedBlockArray + sizeof(uint8*)));
             uint32 mod = shift % blockSize;
             if (mod == 0)
             {
